feat(player): rename 'spatial-navigation-keycode.js' filename
