+++ src/js/utils/spatialkeys.js
